pyproject.toml
[tool.poetry]
name = "forsit-backend-project"
version = "0.1.0"
description = ""
authors = ["Tayyab athar <tayyab.dev.acc@gmail.com>"]
readme = "README.md"
packages = [{include = "forsit_backend_project"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.8"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
flake8-module-name = "^0.2.0"
pep8-naming = "^0.13.3"
pre-commit = "^3.4.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
